+++ server/src/main/java/com/FreeKik/server/MatchController.java
@@ -1,4 +1,4 @@
-package com.FreeKik.server.Controllers;
+package com.FreeKik.server;
 
 import com.FreeKik.server.Handlers.MatchHandler;
 import com.FreeKik.server.models.Match;
